Fixed compilation issues with Microsoft Visual Studio 2010 IDE.
It's strange, but according to compilation errors, it seems that function os_control.c:fake_goal is called only on Windows. Is it correct?

NOTE: in MSVC the keyword inline should be used for C++ files while __inline should be used for C files. I've added a define in order to not modify several files.