The old tiny resolutions has been removed from the selection menu (they can be restored defining OLD_RES costant).
The new resolutions are: 640x480, 800x600, 1024x768 and 1280x1024, there is also the initial 480x320 defined in the etw.cfg config file. As soon this file could be updated, this resolution could be removed. 

NOTE: does is scaling still used? Because any selection it doesn't seems to affect the game so if it's an old featured needed on Amiga it could be removed from the menu.